Removed Short timestamp option as a dumb idea

+++ app/models/card.rb
@@ -1,6 +1,6 @@
 class Card
   include Mongoid::Document
-  include Mongoid::Timestamps::Short
+  include Mongoid::Timestamps
   include Domainable
   include Sluggable
   include Authable
