+++ app/models/ticket.rb
@@ -1,6 +1,6 @@
 class Ticket
   include Mongoid::Document
-  include Mongoid::Timestamps::Short
+  include Mongoid::Timestamps
   include Domainable
   include Sluggable
 end
